+++ packages/apptive_grid_core/lib/src/apptive_grid.dart
@@ -73,8 +73,7 @@ class _ApptiveGridState extends State<ApptiveGrid> {
     super.didUpdateWidget(oldWidget);
     if (oldWidget.options != widget.options) {
       setState(() {
-        _client.options = widget.options;
-        _client.updateEnvironment(widget.options.environment);
+        _client.setOptions(widget.options);
       });
     }
   }