+++ packages/apptive_grid_core/CHANGELOG.md
@@ -1,5 +1,10 @@
+## 0.14.0
+**BREAKING CHANGE**
+- Restructure `ApptiveGridAuthenticator` to handle changes in options correctly by using a specific `ApptiveGridClient` inside the authenticator
+- Update dependencies
+
 ## 0.13.4+1
-* Fix racecondition when refreshing the access token
+* Fix race condition when refreshing the access token
 
 ## 0.13.4
 * Added `DataType.email`, `DataType.phoneNumber` and `DataType.signature`